remove replaced views and VMs
