+++ UnchainedLauncher.GUI/UnchainedLauncher.GUI.csproj
@@ -23,6 +23,7 @@
     <AssemblyName>UnchainedLauncher</AssemblyName>
     <RootNamespace>UnchainedLauncher.GUI</RootNamespace>
     <Configurations>Debug;Release;Debug-FakeChivalryLaunch</Configurations>
+	<XamlDebuggingInformation>True</XamlDebuggingInformation>
   </PropertyGroup>
   <ItemGroup>
     <None Remove="DefaultMaps.txt" />
