+++ UnchainedLauncher.GUI/ViewModels/MainWindowVM.cs
@@ -7,28 +7,21 @@ namespace UnchainedLauncher.GUI.ViewModels {
         public LauncherVM LauncherViewModel { get; }
         public ModListVM ModListViewModel { get; }
         public SettingsVM SettingsViewModel { get; }
-        public ServerLauncherVM ServerLauncherViewModel { get; }
-        public ServersVM ServersViewModel { get; }
         public ServersTabVM ServersTab { get; }
 
         public MainWindowVM(LauncherVM launcherViewModel,
                             ModListVM modListViewModel,
                             SettingsVM settingsViewModel,
-                            ServerLauncherVM serverLauncherViewModel,
-                            ServersVM serversViewModel,
                             ServersTabVM serversTab) {
             LauncherViewModel = launcherViewModel;
             ModListViewModel = modListViewModel;
             SettingsViewModel = settingsViewModel;
-            ServerLauncherViewModel = serverLauncherViewModel;
-            ServersViewModel = serversViewModel;
             ServersTab = serversTab;
         }
 
         public void Dispose() {
             SettingsViewModel.Dispose();
-            ServerLauncherViewModel.Dispose();
-            ServersViewModel.Dispose();
+            // TODO: Servers tab needs to be disposable
         }
     }
 }
