+++ UnchainedLauncher.GUI/Views/MainWindow.xaml
@@ -41,17 +41,11 @@
         <TabItem Header="Mod Manager">
             <local:ModList DataContext="{Binding ModListViewModel}" />
         </TabItem>
-        <TabItem Header="Server">
-            <local:ServerLauncher DataContext="{Binding ServerLauncherViewModel}" />
-        </TabItem>
-        <TabItem Header="NEW Servers TAB">
+        <TabItem Header="Servers">
             <local:ServersTab DataContext="{Binding ServersTab}" />
         </TabItem>
         <TabItem Header="Settings">
             <local:Settings DataContext="{Binding SettingsViewModel}" />
-        </TabItem>
-        <TabItem Header="Running Servers">
-            <local:Servers DataContext="{Binding ServersViewModel}" />
         </TabItem>
     </TabControl>
