--- UnchainedLauncher.GUI/Views/ServerLauncher.xaml
@@ -1,68 +1,0 @@
-﻿<UserControl x:Class="UnchainedLauncher.GUI.Views.ServerLauncher"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
-             xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
-             xmlns:local="clr-namespace:UnchainedLauncher.GUI.Views"
-             xmlns:vms="clr-namespace:UnchainedLauncher.GUI.ViewModels"
-             xmlns:instances="clr-namespace:UnchainedLauncher.GUI.Views.DesignInstances"
-             mc:Ignorable="d" 
-             d:DataContext="{x:Static instances:ServerLauncherViewModelInstances.DEFAULT}"
-             d:DesignHeight="450" d:DesignWidth="800">
-
-    <Grid HorizontalAlignment="Stretch" Margin="0,0,0,0" Width="NaN" VerticalAlignment="Stretch">
-        <Grid.RowDefinitions>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="*"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-            <RowDefinition Height="Auto"/>
-        </Grid.RowDefinitions>
-
-        <Grid.ColumnDefinitions>
-            <ColumnDefinition Width="*"/>
-            <ColumnDefinition Width="*"/>
-            <ColumnDefinition Width="*"/>
-            <ColumnDefinition Width="*"/>
-        </Grid.ColumnDefinitions>
-
-        <TextBlock Text="Server Name :" Grid.Row="0" Grid.ColumnSpan ="2" HorizontalAlignment="Left" VerticalAlignment="Center" Margin="5" />
-        <TextBox Text="{Binding ServerName}" Grid.Row="1" Grid.ColumnSpan ="2" TextWrapping="Wrap" Margin ="5,0,5,0" IsManipulationEnabled="True" AcceptsReturn="False" />
-
-        <TextBlock Text="Password:" Grid.Row="0" Grid.Column="2" Grid.ColumnSpan="1" HorizontalAlignment="Left" VerticalAlignment="Center" Margin="5" />
-        <TextBox Text="{Binding ServerPassword}" Grid.Row="1" Grid.Column ="2" TextWrapping="Wrap" Margin ="5,0,5,0" IsManipulationEnabled="True" AcceptsReturn="False" />
-
-        <TextBlock Text="Select Starting Map:" Grid.Row="0" Grid.Column="3" Grid.ColumnSpan="1" HorizontalAlignment="Left" VerticalAlignment="Center" Margin="5" />
-        <ComboBox ItemsSource="{Binding MapsList, Mode=TwoWay}" SelectedItem="{Binding SelectedMap, Mode=TwoWay}"  Grid.Row="1" Grid.Column="3" Grid.ColumnSpan ="1" Margin ="5,0,5,0" IsManipulationEnabled="True" />
-
-        <TextBlock Text="Server Description:" Grid.Row="2" Grid.ColumnSpan ="4" HorizontalAlignment="Left" VerticalAlignment="Center" Margin="5"/>
-        <TextBox Text="{Binding ServerDescription}" Grid.Row="3" Grid.ColumnSpan ="4" TextWrapping="Wrap" IsManipulationEnabled="True" AcceptsReturn="True" />
-
-        <CheckBox IsChecked="{Binding ShowInServerBrowser}" Grid.Column="2" Grid.Row="2" Grid.ColumnSpan ="2" Content="Show in Server Browser" Margin="5" Height="NaN" VerticalAlignment="Stretch" Width="NaN" HorizontalAlignment="Right" FlowDirection="RightToLeft" />
-
-        <TextBlock Text="Game port:" Grid.Row="4" Grid.Column = "0" HorizontalAlignment="Left" VerticalAlignment="Center" Margin="5"/>
-        <TextBox Text="{Binding GamePort}" Grid.Row="5" Grid.Column = "0" TextWrapping="Wrap" Margin ="5" IsManipulationEnabled="True" AcceptsReturn="True" />
-
-        <TextBlock Text="RCON port:" Grid.Row="4" Grid.Column = "1" HorizontalAlignment="Left" VerticalAlignment="Center" Margin="5"/>
-        <TextBox Text="{Binding RconPort}" Grid.Row="5" Grid.Column = "1" TextWrapping="Wrap" Margin ="5" IsManipulationEnabled="True" AcceptsReturn="True" />
-
-        <TextBlock Text="A2S port:" Grid.Row="4" Grid.Column = "2" HorizontalAlignment="Left" VerticalAlignment="Center" Margin="5"/>
-        <TextBox Text="{Binding A2sPort}" Grid.Row="5" Grid.Column = "2" TextWrapping="Wrap" Margin ="5" IsManipulationEnabled="False" AcceptsReturn="True" />
-
-        <TextBlock Text="Ping port:" Grid.Row="4" Grid.Column = "3" HorizontalAlignment="Left" VerticalAlignment="Center" Margin="5"/>
-        <TextBox Text="{Binding PingPort}" Grid.Row="5" Grid.Column = "3" TextWrapping="Wrap" Margin ="5" IsManipulationEnabled="False" AcceptsReturn="True" />
-
-        <Button Grid.Column="0" Grid.Row="6" Grid.ColumnSpan="2" 
-                Content="Launch Server"  
-                ToolTip="{Binding ButtonToolTip}"
-                Command="{Binding LaunchServerCommand}" 
-                IsEnabled="{Binding Settings.CanClick}" />
-        <Button Grid.Column="2" Grid.Row="6" Grid.ColumnSpan="2" 
-                Content="Launch Headless Server" 
-                ToolTip="{Binding ButtonToolTip}"
-                Command="{Binding LaunchServerHeadlessCommand}" 
-                IsEnabled="{Binding Settings.CanClick}"/>
-    </Grid>
-</UserControl>
