--- UnchainedLauncher.GUI/Views/ServerLauncher.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace UnchainedLauncher.GUI.Views {
-    /// <summary>
-    /// Interaction logic for ServerLauncherView.xaml
-    /// </summary>
-    public partial class ServerLauncher : UserControl {
-        public ServerLauncher() {
-            InitializeComponent();
-        }
-    }
-}
