--- UnchainedLauncher.GUI/Views/ServersTab/Servers.xaml.cs
@@ -1,12 +1,0 @@
-﻿using System.Windows.Controls;
-
-namespace UnchainedLauncher.GUI.Views {
-    /// <summary>
-    /// Interaction logic for ServersView.xaml
-    /// </summary>
-    public partial class Servers : UserControl {
-        public Servers() {
-            InitializeComponent();
-        }
-    }
-}
